+++ builder_pkgs/mockito/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 5.3.3-dev
 
 * Require analyzer 5.2.0.
+* Fix nice mocks generation in mixed mode (generated code is pre null-safety,
+  while mocked class is null-safe).
 * Require Dart >= 2.17.0.
 
 ## 5.3.2